Basics of widget manager system
Allows a function for defining a widget.
Basically it's just a function that gets passed the element of the form
field and the password scoring object. It takes it from there.
widgetManagers need to even set up their own event listeners (but they
don't need to call test() from the scoring object, that's already taken
care of.